[moto] mach-omap2: PM, voltage, CPU:  Revert most PM related drivers back to Moto stock based drivers Voltage changes should avoid device heating up (was a mistake in hynrid driver usage) PM changes gives better battery life while keeping great performance This fixes thermal stuff as well as compatibility to some older governors Fixed the duty cycle cooling function, now more useful for OCers

+++ arch/arm/plat-omap/include/plat/clock.h
@@ -88,7 +88,7 @@ struct clkops {
 struct clksel_rate {
 	u32			val;
 	u8			div;
-	u16			flags;
+	u8			flags;
 };
 
 /**
